appointment agenda opacity
the touchable opacity was altered so that the color of the button changes according to the appointmentType that is received in the api.

I noticed that the the time slots are not in order of time. so i need to fix it so that it looks like a more real calendar