auth
Implemented the clerk auth package and created a new index.tsx page which is where the user first lands when opening app and has buttons to route to the new sign-in and up pages.

the initial page when user logs in is now called home and will appear after user signs in.