auth flow
the authentication is working for sign up and sign in.

now a user must sign up to view data but it is attached to the admin still.

need to create a registration flow so that the new user can input their profile info and need to create a sign out button