onboarding flow
the onboarding flow is complete. a user can now sign up and after inputting the verification code they are redirected to the onboarding screen which they only fill out once and then routes them to the home page.

if the user signs out they are not promted to fill out the onboarding form again.

now i need to make sure appointments can be assigned to them.