toast provider and authentication
when the onboarding is complete the  user now has access to their homepage which now lists appointments.

need to make some adjustments to make only the appointments that were assigned to the interpreter are visible as they can see all appointments including those that are not assigned to them.

Toast provider was added to root layout so now it works as intended