time label and handleSubmit
a new time lable state was added to the time picker component. now after the user selects a time, the time displays inside the button rather than the plain text of select a time.

applied the handle submit to the submit button, but still need to find out why it is not submitted