map and external navigation
the map component now renders on the id page based on the logitude and latitude of the current appointment which is based on the location of the facility. when the user clicks on the link under the map which displays the address. the user is routed to their native map to navigate from their start location to the facility location.

the facility address details are pulled from the api where the admin creates the facility